requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-core
langchain-community
pydantic
pdfplumber
python-dotenv
langchain-groq
